Fixes list formatting and img for spinach lasagna
The list formatting for the spinach lasagna recipe was incorrect.
This resulted in each element in the Directions and Ingredients sections
being its own a separate list, without bullets or numbers.

This commit updates each element to be part a larger list—either the
unordered Ingredients list or the ordered Directions list.

The commit also updates the image source to the correct file name,
as previously it wouldn't display in the browser.