add sweet potato and cauliflower images + links
This commit adds the images for the sweet potato and cauliflower pasta
pages to the recipes folder so they can be accessed later. It also adds
the recipe pages as links to the homepage and, to make the
homepage look cleaner, it reformats all the links as an unordered list.
Finally, it fixes the title in the spicy cauliflower boilerplate.